Extract shared CraftingQuotePanel — dedupe the two crafting dialogs (#2240)
CreateItemDialog had copied AttachFacetDialog's quote panel (cost/skill-cap/
failure-risk + LabStation status), tripping SonarCloud's new-code duplication
gate (11.5% vs 3%). Extracted it to one CraftingQuotePanel component both
dialogs import — less code, no duplication. AttachFacetDialog's 14 tests still
pass (the crafting-quote-panel testid is preserved).
