+++ frontend/src/inventory/components/AttachFacetDialog.tsx
@@ -30,88 +30,7 @@ import {
   useItemFacets,
   useRemoveItemFacet,
 } from '../hooks/useItemFacets';
-import type { CraftingQuote } from '../api';
-import { LabStationStatusCard } from './LabStationStatusCard';
-
-// ---------------------------------------------------------------------------
-// Failure-risk consumption labels
-// ---------------------------------------------------------------------------
-
-const CONSUMPTION_LABELS: Record<string, string> = {
-  none: 'nothing lost',
-  partial: 'some materials/effort lost',
-  full: 'all materials & effort lost',
-};
-
-// ---------------------------------------------------------------------------
-// CraftingQuotePanel — cost / cap / risk summary for a facet selection
-// ---------------------------------------------------------------------------
-
-interface CraftingQuotePanelProps {
-  isLoading: boolean;
-  quote: CraftingQuote | undefined;
-  /** Forwarded to `LabStationStatusCard` — see its `onRepaired` prop. */
-  onStationRepaired?: () => void;
-}
-
-function CraftingQuotePanel({ isLoading, quote, onStationRepaired }: CraftingQuotePanelProps) {
-  if (isLoading) {
-    return <p className="text-xs text-muted-foreground">Loading cost estimate…</p>;
-  }
-  if (!quote) return null;
-
-  const { costs, affordable, max_quality_tier, failure_risk } = quote;
-
-  // Build a cost summary line, omitting zero-cost vectors.
-  const costParts: string[] = [];
-  if (costs.action_points > 0) {
-    costParts.push(`${costs.action_points_have}/${costs.action_points} AP`);
-  }
-  if (costs.anima > 0) {
-    costParts.push(`${costs.anima_have}/${costs.anima} Anima`);
-  }
-  for (const mat of costs.materials) {
-    costParts.push(`${mat.have}/${mat.quantity_required} ${mat.name}`);
-  }
-
-  return (
-    <div
-      className="space-y-1 rounded-md border bg-muted/40 px-3 py-2 text-sm"
-      data-testid="crafting-quote-panel"
-    >
-      {costParts.length > 0 && (
-        <p>
-          <span className="font-medium">Costs:</span>{' '}
-          <span className={affordable ? '' : 'text-destructive'}>{costParts.join(' · ')}</span>
-        </p>
-      )}
-
-      {max_quality_tier != null ? (
-        <p>
-          <span className="font-medium">Skill cap:</span> {max_quality_tier.name}
-        </p>
-      ) : (
-        <p className="text-muted-foreground">No skill cap</p>
-      )}
-
-      {failure_risk.length > 0 && (
-        <p>
-          <span className="font-medium">On failure:</span>{' '}
-          {failure_risk
-            .map((r) => r.label ?? CONSUMPTION_LABELS[r.cost_consumption] ?? r.cost_consumption)
-            .join(', ')}
-        </p>
-      )}
-
-      {quote.station_status && (
-        <LabStationStatusCard
-          featureInstanceId={quote.station_status.feature_instance_id}
-          onRepaired={onStationRepaired}
-        />
-      )}
-    </div>
-  );
-}
+import { CraftingQuotePanel } from './CraftingQuotePanel';
 
 interface AttachFacetDialogProps {
   open: boolean;
